CSV export mass refactoring
