+++ apps/block_scout_web/lib/block_scout_web/controllers/address_transaction_controller.ex
@@ -10,27 +10,13 @@ defmodule BlockScoutWeb.AddressTransactionController do
   import BlockScoutWeb.Models.GetAddressTags, only: [get_address_tags: 2]
   import Explorer.Chain.SmartContract, only: [burn_address_hash_string: 0]
 
-  alias BlockScoutWeb.{AccessHelper, CaptchaHelper, Controller, TransactionView}
-  alias BlockScoutWeb.API.V2.{ApiView, CSVExportController}
+  alias BlockScoutWeb.{AccessHelper, Controller, TransactionView}
   alias Explorer.{Chain, Market}
-  alias Explorer.Chain.Address
-
-  alias Explorer.Chain.CSVExport.{
-    AddressInternalTransactionCsvExporter,
-    AddressLogCsvExporter,
-    AddressTokenTransferCsvExporter,
-    AddressTransactionCsvExporter
-  }
-
-  alias Explorer.Chain.CSVExport.Celo.AddressElectionRewardsCsvExporter,
-    as: CeloAddressElectionRewardsCsvExporter
 
   alias Explorer.Chain.{DenormalizationHelper, Transaction, Wei}
 
   alias Indexer.Fetcher.OnDemand.CoinBalance, as: CoinBalanceOnDemand
   alias Phoenix.View
-
-  alias Plug.Conn
 
   @transaction_necessity_by_association [
     necessity_by_association: %{
@@ -168,70 +154,5 @@ defmodule BlockScoutWeb.AddressTransactionController do
             not_found(conn)
         end
     end
-  end
-
-  defp items_csv(
-         conn,
-         %{
-           "address_id" => address_hash_string,
-           "from_period" => from_period,
-           "to_period" => to_period
-         } = params,
-         csv_export_module
-       )
-       when is_binary(address_hash_string) do
-    with {:ok, address_hash} <- Chain.string_to_address_hash(address_hash_string),
-         {:address_exists, true} <- {:address_exists, Address.address_exists?(address_hash)},
-         {:recaptcha, true} <- {:recaptcha, CaptchaHelper.recaptcha_passed?(params)} do
-      filter_type = Map.get(params, "filter_type")
-      filter_value = Map.get(params, "filter_value")
-
-      address_hash
-      |> csv_export_module.export(from_period, to_period, filter_type, filter_value)
-      |> Enum.reduce_while(CSVExportController.put_resp_params(conn), fn chunk, conn ->
-        case Conn.chunk(conn, chunk) do
-          {:ok, conn} ->
-            {:cont, conn}
-
-          {:error, :closed} ->
-            {:halt, conn}
-        end
-      end)
-    else
-      :error ->
-        unprocessable_entity(conn)
-
-      {:address_exists, false} ->
-        not_found(conn)
-
-      {:recaptcha, false} ->
-        conn
-        |> put_status(:forbidden)
-        |> put_view(ApiView)
-        |> render(:message, %{message: "Invalid reCAPTCHA response"})
-    end
-  end
-
-  defp items_csv(conn, _, _), do: not_found(conn)
-
-  def token_transfers_csv(conn, params) do
-    items_csv(conn, params, AddressTokenTransferCsvExporter)
-  end
-
-  def transactions_csv(conn, params) do
-    items_csv(conn, params, AddressTransactionCsvExporter)
-  end
-
-  def internal_transactions_csv(conn, params) do
-    items_csv(conn, params, AddressInternalTransactionCsvExporter)
-  end
-
-  def logs_csv(conn, params) do
-    items_csv(conn, params, AddressLogCsvExporter)
-  end
-
-  @spec celo_election_rewards_csv(Conn.t(), map()) :: Conn.t()
-  def celo_election_rewards_csv(conn, params) do
-    items_csv(conn, params, CeloAddressElectionRewardsCsvExporter)
   end
 end
