+++ apps/block_scout_web/lib/block_scout_web/routers/tokens_api_v2_router.ex
@@ -69,7 +69,7 @@ defmodule BlockScoutWeb.Routers.TokensApiV2Router do
     get("/:address_hash_param/counters", V2.TokenController, :counters)
     get("/:address_hash_param/transfers", V2.TokenController, :transfers)
     get("/:address_hash_param/holders", V2.TokenController, :holders)
-    get("/:address_hash_param/holders/csv", V2.CSVExportController, :export_token_holders)
+    get("/:address_hash_param/holders/csv", V2.CsvExportController, :export_token_holders)
     get("/:address_hash_param/instances", V2.TokenController, :instances)
     get("/:address_hash_param/instances/:token_id", V2.TokenController, :instance)
     get("/:address_hash_param/instances/:token_id/transfers", V2.TokenController, :transfers_by_instance)
