+++ apps/explorer/test/explorer/chain/csv_export/address/logs_test.exs
@@ -1,8 +1,8 @@
-defmodule Explorer.Chain.AddressLogCsvExporterTest do
+defmodule Explorer.Chain.Address.LogsTest do
   use Explorer.DataCase
 
   alias Explorer.Chain.Address
-  alias Explorer.Chain.CSVExport.AddressLogCsvExporter
+  alias Explorer.Chain.CsvExport.Address.Logs
 
   @first_topic_hex_string_1 "0x7fcf532c15f0a6db0bd6d0e038bea71d30d808c7d98cb3bf7268a95bf5081b65"
   @second_topic_hex_string_1 "0x00000000000000000000000098a9dc37d3650b5b30d6c12789b3881ee0b70c16"
